[Refactor] Move ArrayColumnWriter out of ColumnWriter
